Update
- Added gold pickup
- Added temp Gold UI

Current bugs
- No synced animation

+++ Miner Wars/Assets/Gold.cs
@@ -4,15 +4,7 @@ using UnityEngine;
 
 public class Gold : MonoBehaviour
 {
-
-    public float goldWorth;
-
-
-
-
-
-
-
+    public int goldWorth;
 
 
 }
